[python-frontend] move UTF-8 helpers and chr/ord handlers to function_call/str_conv.cpp
